Allow specifying over which protocols the download can happen
The hawkbit ddi interface has both "download" and "download-http" fields
for _links in deployments as well as matching md5sum fiels. The former
being specified as meant for https the latter for plain http.

Add support for specifying over which protocols a chunk is exposed to be
able to test various situations. Note that this doesn't actually cause a
chunk to be exposed over https, just that the fields meant for https get
used.

Signed-off-by: Sjoerd Simons <sjoerd@collabora.com>