Improve deserialisation of _links map in artifacts
The _links map of an artifact can contain either download, download-http
or both (as well as the matching md5sum fields). Add a custom
deserializer to cope with that as well as splitting up the Links struct
in substructs per protocol type.

The deserializer enforces either download or download_http exists so
there should always be at least one download link. The download helpers
prefer the https (download) url if it exists.

Signed-off-by: Sjoerd Simons <sjoerd@collabora.com>