feat: 'use-rustls' and 'native-tls' features
This commit introduces two new features:
- (default) `native-tls` which enables the reqwest `default-tls`
feature -> this configuration is the same as the previous
default. Therefore no change in default behvior is introduced.
- `use-rustls` replaces the `default-tls` feature with the
`rustls-tls-manual-roots` reqwest feature. This removes the openssl
dependency and uses the rustls ssl implemention.

This change is useful, because for some use cases it may not be
optimal to depend on the openssl system lib. For example cross-compilation.